Deleted CHANGELOG line not belonging to patch 2
Note: This is an exception to the rule that 'tags' sub-directories should not
be edited. It is too convenient not to do it in this case, because we are
trying to re-create a historic state of the repository for the sake of being
able to check it out with a symbolic tag name later.


git-svn-id: http://svn.freetz.org/tags/14.1-p2@139 f5190166-0702-4917-9039-51ec32eddaf5