uClibc:  - remove unnecessary UCLIBC_EXTRA_CFLAGS and UCLIBC_EXTRA_LDFLAGS in uClibc-install rule.  - remove | sign as it's a real dependency and not the order-only one.  - refs #1049
The flags are unnecessary because uClibc is already compiled, we just wanna install it. So passing them has absolutely no sense. Moreover these flags are not the same as those used to compile uClibc. Starting with version 0.9.30 uClibc has a dependency tracking system being able to recognize compile flags changes. Thus passing them causes uClibc to be recompiled, which in turn causes gcc-final to be recompiled.


git-svn-id: http://svn.freetz.org/trunk@5861 f5190166-0702-4917-9039-51ec32eddaf5