[PATCH] netfilter/IPv6: Backport ip6t_REJECT from linux-2.6.14 (refs #1394)  * by dileks
While playing with IPv6 and iptables to protect my Internet connection,
I tried to follow the SixXS "IPv6 stateless firewall" example-script [1].
Unfortunately, ip6t_REJECT kernel-module is required to run this script
successfully.
My W701V is based on Linux-2.6.13.1 and has no support for this feature.

REJECT target support for netfilter/IPv6 was first introduced in 2.6.14.
So, this patch includes a backport of it for Linux-2.6.13.1.

Furthermore, CONFIG_IP6_NF_TARGET_REJECT option is enabled as module for
all boxes with AR7 and OHIO kernel-layout as requested by Oliver on IRC.

Tested with 04.80 and 04.87 on a Speedport W701V router.

See also <http://freetz.org/ticket/1394>.

git-svn-id: http://svn.freetz.org/trunk@7213 f5190166-0702-4917-9039-51ec32eddaf5