Package `ministun`: Change internal structure analogous to `debootstrap`, i.e. use macros `PKG_LOCALSOURCE_PACKAGE` and `PKG_CONFIGURED_NOP` instead of `PKG_UNPACKED` and `src` package subdirectory instead of `patches`. Thus, we have a directly editable source code file instead of a patch.
Thanks to er13 for pointing out this possibility.

Note: This package will probably be removed again from the repository soon
because ralf's new `stunclient` applet for Busybox is even smaller and due to
be used by `get_ip`. But before this package goes away, I wanted to fix it for
later reference, if we ever need it (or something similar) again.


git-svn-id: http://svn.freetz.org/trunk@8254 f5190166-0702-4917-9039-51ec32eddaf5