Revert r8570 because it was right before and wrong afterwards. Sorry Whoopie.
'''Edit:''' My version works when the script is called stand-alone, yours only
worked when it was sourced by ''fwmod''. In a way both versions are wrong as in
"not universal". I have an idea how to solve this, but the workaround is in
r8573 (sorry for the strange forward reference).

git-svn-id: http://svn.freetz.org/trunk@8572 f5190166-0702-4917-9039-51ec32eddaf5