Add subshell support for BusyBox Ash, i.e. Bash package is no longer needed.
Hint for Bash users: Bash will not be automatically chosen for the subshell
anymore even if present in the firmware. If you want Bash, you now really have
to use it as your login shell as usual in MC, i.e. you need to
  - choose `FREETZ_PACKAGE_BASH` (like before),
  - choose sub-option `FREETZ_PATCH_BASH_LOGIN_SHELL`,
  - edit ''/etc/passwd'' on the box for the user who wants Bash as her
    login shell and MC subshell.

See comments in ''patches/040-ash_as_subshell.patch'' for more details.


git-svn-id: http://svn.freetz.org/trunk@8754 f5190166-0702-4917-9039-51ec32eddaf5