On Ubuntu Linux systems the default umask settings (022) has changed. More precisely the umask line was commented in /etc/profile.
Some files on the rootfs in Freetz require 755 as file permissions
to run correctly.
For example files located in make/mod/files/root/.
(Not sure if other files needs to be fixed, too.)

The displayed error-message with hint to fix the umask problem were
not really helpful (seen on Ubuntu/precise).
Try to find a better wording.

The below reference could be included in the Freetz FAQ.

From [1]:

  In /etc/profile please comment:

  # umask 022
  # Setting a fixed umask in shell configuration files was only a workaround.
  # It is no longer needed and should not be done since the umask can be
  # set consistently with PAM. (See man pam_umask.)

Hint: Check your umask settings in /etc/profile and/or ~/.profile!

[1] https://lists.ubuntu.com/archives/foundations-bugs/2011-December/056446.html

Signed-off-by: Sedat Dilek <sedat.dilek@gmail.com>

git-svn-id: http://svn.freetz.org/trunk@9427 f5190166-0702-4917-9039-51ec32eddaf5