7140 (English, AnnexA & AnnexB):  * bump firmware version to 39/30.04.77 rev27510 ([http://www.avm.de/de/Sicherheit/hinweis.html security update])  * compile-tested only  * refs #2355
TODO: removed patches/devices/04_XX/7140/en/210-notmountable.patch because it doesn't exist in AT_CH version (the patched file is identical in both versions). It could however be the case that AT_CH is just missing the patch - check it and readd the patch if necessary to both versions, i.e. AT_CH & EN. ditto for r11783.


git-svn-id: http://svn.freetz.org/trunk@11782 f5190166-0702-4917-9039-51ec32eddaf5