yourfritz-host: make signed images reproducible
I.e. signing the same input file is expected to produce the same output file.
Up to now this was not the case as the ./var/signature member had the signing
time as its timestamp which in turn led to different output files.

Fix it by determining the timestamp of the newest member in the image and using
it also for the ./var/signature member.


git-svn-id: http://svn.freetz.org/trunk@14841 f5190166-0702-4917-9039-51ec32eddaf5