kconfig: update version numbers and source of code
Using 'repo.or.cz' (a mirror of the linux kernel git repo) without any
validation method is - in my opinion - INSECURE ... no-one would
recognize, if someone tampered with the files from such an unimportant
sub-directory like 'scripts' and it's not clear, who has write access to
repo.or.cz.

But I've no better idea and see no obvious way yet, to verify the
integrity of files from 'git-archive', if the automatic extraction takes
place.

Let's hope, that at least one of the mirrors is available and will
provide the pre-processed compressed archive file; the mirrors are
searched prior to the attempt to use 'git-archive' (in freetz_download).

I would prefer a form of archive validation, even if the file is loaded
from a mirror - but Freetz does not provide any signatures on these
archives yet. But in contrast to packages under "make", here even a
simple hash comparision is missing.

That's why I would suggest to store the source files in a local sub-dir
and to maintain an own copy. It ensures at the same time, that further
updates to the kernel sources are ignored. If kernel sources change and
there's no mirror available at build time (no matter what's the reason
for), the latest version from kernel git gets copied. I'm unsure, if
that's always the intended case.