Add support for using image filenames other than that used by AVM
Starting with Fritz!OS 07.10 AVM changed the naming scheme for the image files.
The filenames are not unique anymore as it used to be, only the full URLs are still unique.

Example:
 Both German and international images for 7530 are called `FRITZ.Box_7530-07.10.image`,
 see http://ftp.avm.de/fritzbox/fritzbox-7530/deutschland/fritz.os/FRITZ.Box_7530-07.10.image
 and http://ftp.avm.de/fritzbox/fritzbox-7530/other/fritz.os/FRITZ.Box_7530-07.10.image
 The files are however not identical.

This change is not compatible with `make` expecting every file to be unique.

To overcome the limitations caused by this change it is now possible to provide
two filenames: the local one (this is expected to be unique) and the remote one
(that used by AVM). The downloaded remote file will be renamed to the unique
local one. In order to use this feature `FREETZ_DL_SOURCE` has to be defined
in the following way `$(LOCAL_UNIQUE_IMAGE_FILENAME)/$(REMOTE_IMAGE_FILENAME)`,
i.e. the local filename followed by the forward slash (the separator) and by
the remote filename.