kernel build process: rework and update kernel-header-symlinks related code.
It has not been updated for a very long time and became incomplete.
The side effect of this incompleteness is that it was not possible
to use delta-patch based method for integrating kernel sources from
AVM opensrc packages. The latter because it's not possible to create
symlinks using the standard patch tool.

The new code has been tested for the following boxes/firmware/kernel versions:
 7170.04.86   (2.6.13)
 7270v1.04.86 (2.6.19)
 7270v3.05.23 (2.6.32)
 7270v3.06.05 (2.6.32)
 7390.05.23   (2.6.28)
 7390.06.30   (2.6.28)
 7390.06.85   (2.6.28)
 7490.06.30   (2.6.32)
 7490.07.11   (3.10)