strace: workaround SO_PROTOCOL/SO_DOMAIN related problem affecting IKS boxes since Fritz!OS-6.5x
AVM backported some code from later kernel versions to 2.6.28, in particular the following #define's:
  #define SO_PROTOCOL 38  /* Backport 2.6.32 */
  #define SO_DOMAIN   39  /* Backport 2.6.32 */

These are valid for other ARCHs but not for MIPS. The proper #define's for MIPS are:
  #define SO_PROTOCOL 0x1028  /* protocol type */
  #define SO_DOMAIN   0x1029  /* domain/socket family */

Unfortunately we cannot fix AVM's backport as the binary versions of the kernels
shipped with Fritz!OS-6.5x and 6.8x are compiled using the wrong codes.

We workaround the problem in user space by making it possible to distinguish
between correct and wrong codes.

Closes https://github.com/Freetz/freetz/issues/208

(cherry picked from commit d54819b3799b81b3892248165d33f2af456882f0)